data module refactoring
    - data/datasets: Each dataset has its own file
    - processing: Check if data given is already numpy
    - requirements: Add Pillow
