+++ .gitignore
@@ -4,7 +4,9 @@
 __pycache__/
 *.pyc
 *.swp
+.cache/
 pypi_upload.sh
+notes/
 
 # Distribution / packaging
 bin/
