+++ requirements.txt
@@ -2,6 +2,7 @@ numpy==1.18.1
 torch==1.4.0
 torchvision==0.5.0
 torchsummary==1.5.1
+Pillow==7.0.0
 matplotlib==3.1.2
 albumentations==0.4.5
 opencv-python==4.1.2.30
