+++ tensornet/data/datasets/__init__.py
@@ -1,0 +1,2 @@
+from .mnist import MNIST
+from .cifar10 import CIFAR10
